Deprecate split view modifiers (#29)
After some experimentation I came to the conclusion that split view
"transitions" are not meant to be modified directly. They're not even a
thing in itself, really.

Instead, it should be the stack navigation view *within* the split view
that can and should have their transition modified directly.

So I deprecated all the split view modifiers and flattened the 2 stack
modifiers into one: `.navigationTransition`. In addition to this, a
useful runtime warning will let you know when you're holding it wrong.

This drastically simplifies the API surface to a single entry point and
is, I believe, a step in the right direction to easing the entry barrier
for this library.